requirements.txt
absl-py==2.1.0
immutabledict==4.2.0
iniconfig==2.0.0
numpy==2.1.2
ortools==9.11.4210
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pluggy==1.5.0
protobuf==5.26.1
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.2
